Make filesize options use byte-precise u64 values and document suffixes
There's currently only 1 such option, but later commits will make
e.g. `SYN_HIGHLIGHT_MAX_LINE_LEN` configurable, which will require
sub-MiB precision.

As mentioned in the `CHANGELOG.md` entry, this is technically a
(silently) breaking change to the `filesize-limit` option, but I'm
justifying this by the fact that it's probably configured by very
few people and the consequences of it being lower than expected
(bytes instead of MiB) are harmless.

+++ src/options.h
@@ -82,7 +82,7 @@ typedef struct {
     uint8_t msg_compile; // Default EditorState::messages[] index for `compile`
     uint8_t msg_tag; // Default EditorState::messages[] index for `tag`
     unsigned int esc_timeout; // See term_read_input()
-    unsigned int filesize_limit; // Size limit imposed by load_buffer() (in MiB)
+    uint_least64_t filesize_limit; // Size limit imposed by load_buffer()
     const char *statusline_left;
     const char *statusline_right;
 } GlobalOptions;
